Dockerize app, fixes #1

+++ library/library_main_app/views.py
@@ -3,12 +3,10 @@ from django.shortcuts import render, get_object_or_404
 from drf_spectacular.types import OpenApiTypes
 from drf_spectacular.utils import extend_schema, OpenApiParameter, OpenApiRequest
 from rest_framework import viewsets, status
-from rest_framework.decorators import action
 from rest_framework.response import Response
 from rest_framework.pagination import PageNumberPagination
-from .models import Book, Author
-from .serializers import BookSerializer, AuthorSerializer
-
+from .models import Book, Author, Genre
+from .serializers import BookSerializer, AuthorSerializer, GenreSerializer, UuidSerializer
 
 # Create your views here.
 class MyPagination(PageNumberPagination):
@@ -30,7 +28,7 @@ class AuthorViewSet(viewsets.ViewSet):
     mypagination_class = MyPagination()
     serializer_class = AuthorSerializer
     @extend_schema(
-        tags=['Get'],
+        tags=['Автор'],
         summary="Получение автора по UUID",
     )
     def retrieve(self, request, pk=None):
@@ -39,7 +37,7 @@ class AuthorViewSet(viewsets.ViewSet):
         return Response(serializer.data)
 
     @extend_schema(
-        tags=['Get'],
+        tags=['Автор'],
         summary="Получение страницы с авторами",
         parameters=[
             OpenApiParameter(name='page', location=OpenApiParameter.QUERY, description='Номер страницы',
@@ -60,7 +58,7 @@ class AuthorViewSet(viewsets.ViewSet):
         return resp
 
     @extend_schema(
-        tags=['Post'],
+        tags=['Автор'],
         summary="Добавление автора",
         request=AuthorSerializer
     )
@@ -72,17 +70,22 @@ class AuthorViewSet(viewsets.ViewSet):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     @extend_schema(
-        tags=['Delete'],
-        summary="Удаление автора по UUID",
+        tags=['Автор'],
+        summary="Удаление автора по UUID"
     )
     def destroy(self, request, pk=None):
-        queryset = Author.objects.all()
-        author = get_object_or_404(queryset, pk=pk)
+        uuid = pk
+        if not uuid:
+            return Response(status=status.HTTP_400_BAD_REQUEST)
+        try:
+            author = Author.objects.get(uuid=uuid)
+        except Author.DoesNotExist:
+            return Response(status=status.HTTP_404_NOT_FOUND)
         author.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
     @extend_schema(
-        tags=['Put'],
+        tags=['Автор'],
         summary="Модификация автора по UUID",
         request=AuthorSerializer,
     )
@@ -100,7 +103,7 @@ class BookViewSet(viewsets.ViewSet):
     serializer_class = BookSerializer
 
     @extend_schema(
-        tags=['Get'],
+        tags=['Книга'],
         summary="Получение книги по UUID",
     )
     def retrieve(self, request, pk=None):
@@ -109,7 +112,7 @@ class BookViewSet(viewsets.ViewSet):
         return Response(serializer.data)
 
     @extend_schema(
-        tags=['Get'],
+        tags=['Книга'],
         summary="Получение страницы с книгами",
         parameters=[
             OpenApiParameter(name='page', location=OpenApiParameter.QUERY, description='Номер страницы',
@@ -130,7 +133,7 @@ class BookViewSet(viewsets.ViewSet):
         return resp
 
     @extend_schema(
-        tags=['Post'],
+        tags=['Книга'],
         summary="Добавление книги",
     )
     def create(self, request):
@@ -141,16 +144,23 @@ class BookViewSet(viewsets.ViewSet):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     @extend_schema(
-        tags=['Delete'],
+        tags=['Книга'],
         summary="Удаление книги по UUID",
     )
     def destroy(self, request, pk=None):
-        book = get_object_or_404(Book, pk=pk)
-        book.delete()
+        uuid = pk
+        if not uuid:
+            return Response(status=status.HTTP_400_BAD_REQUEST)
+        try:
+            author = Book.objects.get(uuid=uuid)
+        except Book.DoesNotExist:
+            return Response(status=status.HTTP_404_NOT_FOUND)
+        author.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
+
     @extend_schema(
-        tags=['Put'],
+        tags=['Книга'],
         summary="Модификация книги по UUID",
         request=BookSerializer
     )
@@ -158,6 +168,81 @@ class BookViewSet(viewsets.ViewSet):
         queryset = Book.objects.all()
         book = get_object_or_404(queryset, pk=pk)
         serializer = BookSerializer(instance=book, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+class GenreViewSet(viewsets.ViewSet):
+    mypagination_class = MyPagination()
+    serializer_class = GenreSerializer
+
+    @extend_schema(
+        tags=['Жанр'],
+        summary="Получение жанра по UUID",
+    )
+    def retrieve(self, request, pk=None):
+        genre = get_object_or_404(Genre, pk=pk)
+        serializer = BookSerializer(genre, many=False)
+        return Response(serializer.data)
+
+    @extend_schema(
+        tags=['Жанр'],
+        summary="Получение жанров с их представителями",
+        parameters=[
+            OpenApiParameter(name='page', location=OpenApiParameter.QUERY, description='Номер страницы',
+                             required=False, type=int),
+            OpenApiParameter(name='name', location=OpenApiParameter.QUERY, description='Поиск по названию жанра',
+                             required=False, type=OpenApiTypes.STR),
+        ],
+    )
+    def list(self, request):
+        queryset = Genre.objects.all().order_by('title')
+        if request.query_params.get('name'):
+            filter_value = request.query_params.get('name')
+            queryset = queryset.filter(Q(title__icontains=filter_value)
+                                       | Q(title__startswith=filter_value))
+        p = self.mypagination_class.paginate_queryset(queryset, request)
+        serializer = GenreSerializer(p, many=True)
+        resp = self.mypagination_class.get_paginated_response(serializer.data)
+        return resp
+
+    @extend_schema(
+        tags=['Жанр'],
+        summary="Добавление жанра",
+    )
+    def create(self, request):
+        serializer = GenreSerializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    @extend_schema(
+        tags=['Жанр'],
+        summary="Удаление жанра по UUID",
+    )
+    def destroy(self, request, pk=None):
+        uuid = pk
+        if not uuid:
+            return Response(status=status.HTTP_400_BAD_REQUEST)
+        try:
+            author = Genre.objects.get(uuid=uuid)
+        except Genre.DoesNotExist:
+            return Response(status=status.HTTP_404_NOT_FOUND)
+        author.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
+
+
+    @extend_schema(
+        tags=['Жанр'],
+        summary="Модификация жанра по UUID",
+        request=BookSerializer
+    )
+    def update(self, request, pk=None):
+        queryset = Genre.objects.all()
+        book = get_object_or_404(queryset, pk=pk)
+        serializer = GenreSerializer(instance=book, data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
